fix(dbt-cloud-api): use int64 types for identifiers (#8311)
* fix(dbt-cloud-api): use int64 types for identifiers

With our method of generating unique numeric ids in the multi-cell
architecture (see
[RFC-87](https://www.notion.so/dbtlabs/RFC-87-Unique-Numeric-IDs-a33f96c518824a40b89ffbf776db576e)),
IDs will no longer fit into 32 bits. Ideally, we would rely on
generated client code entirely, but the dbt-cloud API definition
is lacking the approptiate annotation on the relevant id fields that
would generate the appropriate types for a Rust client. I have
opened dbt-labs/dbt-cloud#16647 to fix OpenAPI generation, but
I manually updated the API types in the meantime.

* chore: add changelog entry

* fix(dbt-cloud-api): replace remaining i32 identifiers with i64

GitOrigin-RevId: c2703ad378bfb8950bce30966f41b7987938b880

+++ crates/dbt-cloud-api/src/apis/credentials_api.rs
@@ -23,8 +23,8 @@ pub enum ListCredentialsError {
 /// List the credentials for a given project.
 pub async fn list_credentials(
     configuration: &configuration::Configuration,
-    account_id: i32,
-    project_id: i32,
+    account_id: i64,
+    project_id: i64,
 ) -> Result<models::CredentialsResponseListEnveloped, Error<ListCredentialsError>> {
     // add a prefix to parameters to efficiently prevent name collisions
     let p_account_id = account_id;
